Integrated API, Updated JS and CSS
Removed the day array previously set. Integrated the API and looped through to display the daily forecast. Removed the icon to replace with live data. Added a function to reduce the days displayed. Changed the CSS to make the Icons look better and the app more readable. Replaced dummy content with real live content.